split functions into modules to make code more readable
