+++ lib/cancan/model_adapters/active_record_adapter.rb
@@ -1,6 +1,9 @@
+require_relative 'can_can/model_adapters/active_record_adapter/joins.rb'
 module CanCan
   module ModelAdapters
     module ActiveRecordAdapter
+      include CanCan::ModelAdapters::ActiveRecordAdapter::Joins
+
       # Returns conditions intended to be used inside a database query. Normally you will not call this
       # method directly, but instead go through ModelAdditions#accessible_by.
       #
@@ -61,16 +64,6 @@ module CanCan
         end
       end
 
-      # Returns the associations used in conditions for the :joins option of a search.
-      # See ModelAdditions#accessible_by
-      def joins
-        joins_hash = {}
-        @rules.each do |rule|
-          merge_joins(joins_hash, rule.associations_hash)
-        end
-        clean_joins(joins_hash) unless joins_hash.empty?
-      end
-
       def database_records
         if override_scope
           @model_class.where(nil).merge(override_scope)
@@ -98,7 +91,7 @@ module CanCan
         rule_found = @rules.detect { |rule| rule.conditions.is_a?(ActiveRecord::Relation) }
         raise Error,
               'Unable to merge an Active Record scope with other conditions. '\
-                "Instead use a hash or SQL for #{rule_found.actions.first} #{rule_found.subjects.first} ability."
+              "Instead use a hash or SQL for #{rule_found.actions.first} #{rule_found.subjects.first} ability."
       end
 
       def merge_conditions(sql, conditions_hash, behavior)
@@ -131,26 +124,6 @@ module CanCan
 
       def sanitize_sql(conditions)
         @model_class.send(:sanitize_sql, conditions)
-      end
-
-      # Takes two hashes and does a deep merge.
-      def merge_joins(base, add)
-        add.each do |name, nested|
-          if base[name].is_a?(Hash)
-            merge_joins(base[name], nested) unless nested.empty?
-          else
-            base[name] = nested
-          end
-        end
-      end
-
-      # Removes empty hashes and moves everything into arrays.
-      def clean_joins(joins_hash)
-        joins = []
-        joins_hash.each do |name, nested|
-          joins << (nested.empty? ? name : { name => clean_joins(nested) })
-        end
-        joins
       end
     end
   end
